Fix narrowing issue

+++ src/structure.cpp
@@ -3483,7 +3483,7 @@ void _syncDebugStructure(const char *function, STRUCTURE const *psStruct, char c
 		psStruct->pos.x, psStruct->pos.y, psStruct->pos.z,
 		(int)psStruct->status,
 		(int)psStruct->pStructureType->type, refChr, ref,
-		psStruct->currentBuildPts,
+		(int)psStruct->currentBuildPts,
 		(int)psStruct->body,
 	};
 	_syncDebugIntList(function, "%c structure%d = p%d;pos(%d,%d,%d),status%d,type%d,%c%.0d,bld%d,body%d", list, ARRAY_SIZE(list));
